Seriously refactored the serialization because it kept throwing errors during the test. Fixed the API endpoints to pass the data correctly. Wrote tests to evaluate the functionality of the endpoints

+++ server/src/main/java/service/UserService.java
@@ -9,7 +9,7 @@ public class UserService {
     private UserDAO userDAO;
     private AuthDAO authDAO;
 
-    public UserService(UserDAO userDAO, AuthDAO authDAO){
+    public UserService(UserDAO userDAO, AuthDAO authDAO) {
         this.userDAO = userDAO;
         this.authDAO = authDAO;
     }
@@ -18,17 +18,23 @@ public class UserService {
         return UUID.randomUUID().toString();
     }
 
-    public AuthData register(UserData user) throws DataAccessException{
+    public AuthData register(UserData user) throws DataAccessException {
         if (userDAO.getUser(user.username()) != null) {
-            throw new DataAccessException("username taken");
+            throw new DataAccessException("already taken");
+        }
+        if (user.username() == null || user.password() == null || user.email() == null) {
+            throw new DataAccessException("bad request");
         }
         userDAO.createUser(user);
         return login(user);
     }
 
-    public AuthData login(UserData user) throws DataAccessException{
+    public AuthData login(UserData user) throws DataAccessException {
+        if (user == null || user.username() == null || user.password() == null) {
+            throw new DataAccessException("bad request");
+        }
         UserData existingUser = userDAO.getUser(user.username());
-        if (existingUser == null) {
+        if (existingUser == null || !existingUser.password().equals(user.password())) {
             throw new DataAccessException("unauthorized");
         }
         AuthData authData = new AuthData(generateToken(), user.username());
@@ -36,7 +42,10 @@ public class UserService {
         return authData;
     }
 
-    public void logout(String authToken) throws DataAccessException{
+    public void logout(String authToken) throws DataAccessException {
+        if (authDAO.getAuth(authToken) == null) {
+            throw new DataAccessException("unauthorized");
+        }
         authDAO.deleteAuth(authToken);
     }
 }
